[rpc] Move DescribeAddressVisitor to rpc/util
