+++ src/Makefile.am
@@ -215,6 +215,7 @@ libbitcoin_server_a_SOURCES = \
   rpc/rawtransaction.cpp \
   rpc/safemode.cpp \
   rpc/server.cpp \
+  rpc/util.cpp \
   script/sigcache.cpp \
   script/ismine.cpp \
   timedata.cpp \
