+++ src/rpc/util.h
@@ -5,6 +5,13 @@
 #ifndef BITCOIN_RPC_UTIL_H
 #define BITCOIN_RPC_UTIL_H
 
+#include <pubkey.h>
+#include <script/standard.h>
+#include <univalue.h>
+#include <utilstrencodings.h>
+
+#include <boost/variant/static_visitor.hpp>
+
 #include <string>
 #include <vector>
 
@@ -15,5 +22,7 @@ class CScript;
 CPubKey HexToPubKey(const std::string& hex_in);
 CPubKey AddrToPubKey(CKeyStore* const keystore, const std::string& addr_in);
 CScript CreateMultisigRedeemscript(const int required, const std::vector<CPubKey>& pubkeys);
+
+UniValue DescribeAddress(const CTxDestination& dest);
 
 #endif // BITCOIN_RPC_UTIL_H
